+++ src/main/java/ru/kelcuprum/sailstatus/presence/AbstractPresence.java
@@ -8,10 +8,9 @@ public abstract class AbstractPresence {
     abstract public void execute();
     abstract public boolean avaliable();
 
-    enum TYPES{
+    public enum TYPES{
         MENU(0),
-        IN_GAME(1),
-        MOD_SUPPORT(2);
+        IN_GAME(1);
         final int type;
         TYPES(int type){
             this.type = type;