+++ src/main/java/ru/kelcuprum/sailstatus/presence/Unknown.java
@@ -4,13 +4,15 @@ import com.jagrosh.discordipc.entities.*;
 import ru.kelcuprum.sailstatus.SailStatus;
 import ru.kelcuprum.sailstatus.config.Assets;
 
+import static ru.kelcuprum.sailstatus.SailStatus.defaultBuilder;
+
 public class Unknown {
     /**
      * Состояние: what?<br>
      * Причины: Ни одна проверка не нашла подходящий вариант
      */
-    public Unknown(){
-        RichPresence.Builder presence = new RichPresence.Builder().setActivityType(ActivityType.Playing);
+    public static void execute(){
+        RichPresence.Builder presence = defaultBuilder();
         presence.setState(SailStatus.localization.getLocalization("unknown", true));
         presence.setLargeImage(Assets.getSelected().getIcon("unknown"));
         SailStatus.sendPresence(presence.build());