+++ src/main/java/ru/kelcuprum/sailstatus/presence/ingame/MultiPlayer.java
@@ -8,9 +8,11 @@ import ru.kelcuprum.sailstatus.SailStatus;
 import ru.kelcuprum.sailstatus.config.Assets;
 import ru.kelcuprum.sailstatus.info.*;
 
+import static ru.kelcuprum.sailstatus.SailStatus.defaultBuilder;
+
 public class MultiPlayer {
-    public MultiPlayer(){
-        RichPresence.Builder presence = new RichPresence.Builder().setActivityType(ActivityType.Playing);
+    public static void execute() {
+        RichPresence.Builder presence = defaultBuilder();
         if(AlinLib.MINECRAFT.getCurrentServer() == null){
             presence.setLargeImage("https://wf.kelcu.ru/mods/waterplayer/icons/seadrive.gif");
             presence.setDetails("🔎 HowTo: fix this shit");
@@ -24,7 +26,7 @@ public class MultiPlayer {
             if(SailStatus.userConfig.getBoolean("ENABLE_WORLD", true)){
                 if(World.getCodeName().endsWith("overworld") && SailStatus.userConfig.getBoolean("ENABLE_TIME_CYCLE", true)) PresenceWorld.getTime(presence);
                 else {
-                    presence.setLargeImage(PresenceWorld.getAssets(), World.getName());
+                    presence.setLargeImage(PresenceWorld.getAssets(), World.getName(), "");
                 }
             } else presence.setLargeImage(Assets.getSelected().getIcon("logo"));
             SailStatus.sendPresence(presence.build());