+++ src/main/java/ru/kelcuprum/sailstatus/presence/ingame/SaveWorld.java
@@ -5,9 +5,11 @@ import com.jagrosh.discordipc.entities.RichPresence;
 import ru.kelcuprum.sailstatus.SailStatus;
 import ru.kelcuprum.sailstatus.config.Assets;
 
+import static ru.kelcuprum.sailstatus.SailStatus.defaultBuilder;
+
 public class SaveWorld {
-    public SaveWorld(){
-        RichPresence.Builder presence = new RichPresence.Builder().setActivityType(ActivityType.Playing);
+    public static void execute() {
+        RichPresence.Builder presence = defaultBuilder();
         presence.setState(SailStatus.localization.getLocalization("save.world", true));
         presence.setLargeImage(Assets.getSelected().getIcon("unknown"));
         SailStatus.sendPresence(presence.build());