+++ changelog.md
@@ -1,2 +1,4 @@
-- Added icon
-- Port to 1.21.9 (25w32a)
+- Code refactor
+- Changed times update
+- Fixes import olds configs
+- Added custom mods presence (SailStatus.registerModPresence(AbstractPresence))